requirements.txt
jinja2==3.1.5
requests==2.32.3
PyYAML==6.0.2
hvac==2.3.0
python-dotenv==1.0.1
